Make index_t an uint64_t
While tinkering with the API I realized there was a huge problem:
Returning RIDs from methods like `body_get_shape` is impossible*, as the
library "forgot" the pointer to the boxed index, which is what the
module used.

The most straightforward fix was to encode the indices into a `u64`
instead of boxing them. By encoding them in a deterministic way, it is
possible to always directly use them for hashmap lookups. It also
reduces memory usage, is safer and is just faster in every way.

The only disadvantage is that it limits the amount of elements per array
to 2^32. I do not anticipate this limit to be reached anytime and if it
does, we can always transition to u128 (even though the ABI for that
isn't well-defined).

* It is technically possible, but it would be cumbersome and slow

+++ native/src/server/body.rs
@@ -31,7 +31,7 @@ enum StateError {
 }
 
 struct Shape {
-	index: usize,
+	index: u32,
 	transform: Isometry<f32>,
 	enabled: bool,
 }
@@ -87,11 +87,11 @@ impl Body {
 		}
 	}
 
-	fn map_shape_mut<F>(&mut self, index: usize, f: F)
+	fn map_shape_mut<F>(&mut self, index: u32, f: F)
 	where
 		F: FnOnce(&mut Shape),
 	{
-		if let Some(shape) = self.shapes.get_mut(index) {
+		if let Some(shape) = self.shapes.get_mut(index as usize) {
 			f(shape);
 		} else {
 			eprintln!("Invalid shape index");
@@ -190,13 +190,13 @@ fn set_shape_transform(
 	shape: i32,
 	transform: &Transform,
 ) {
-	let shape = shape as usize;
+	let shape = shape as u32;
 	map_or_err!(body, map_body_mut, |v| v
 		.map_shape_mut(shape, |v| v.transform = transform_to_isometry(*transform)));
 }
 
 fn set_shape_disabled(body: Index, shape: i32, disable: bool) {
-	let shape = shape as usize;
+	let shape = shape as u32;
 	map_or_err!(body, map_body_mut, |v| v
 		.map_shape_mut(shape, |v| v.enabled = !disable));
 }
